Request permission to send notifications on Android 13
The use of shouldShowRequestPermissionRational() makes Mumla ask for the
POST_NOTIFICATIONS permission 2 times, and then not again. The
mPermPostNotificationsAsked variable makes it so that the 2nd time will
take place after a complete restart of Mumla. Seems good. Noting that I
don't see any "Never ask again" button in the dialog, which I kind of
expected.