Bump to SDK 35 by engaging in major yak shaving...
Forced to raise minSdk to 21 (from 14). So Mumla now requires Android
5.0+ (before Android 4.0.1). This due to some or other newer androidx
library that had to be used for other needed changes, somewhere in this
mess of hair.

Ended up having to overhaul Mumla's theming as part of adapting the UI
layout to edge-to-edge. Its looks different now, but hopefully now
worse; did quite a lot of tweaking to avoid that. System default theme
(light/dark) is now honoured by default, with possibility to force light
or dark in the settings.

The preferences/settings system had to move from some older legacy
system to some, I think, slightly newer legacy system
(androidx.preference). A number of things had to be ported to that.